requirements.txt
fastapi
uvicorn[standard]
Scrapy
chromadb
scrapy-playwright
langchain[llms]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
sentry-sdk[fastapi]
